Filter design and analysis 5.4

+++ Task_5/4_Realization_of_Digital_Filters/FilterDesignAnalysis/FIRFilter.h
@@ -3,7 +3,7 @@
 
 #include <stdint.h>
 
-template<typename Floating, uint8_t ORDER>
+template<typename Floating, uint16_t ORDER>
 class FIRFilter
 {
     public:
